// src/hooks/services/investor/getInvestorList/type.ts
export interface IParams {
  page?: string;
  sortBy?: string;
  sortOrder?: string;
  limit?: string;
  userId?: string;
  accountName?: string;
  accountLevel?: string;
  profileCompletionStatus?: string;
  kycStatus?: string;
  IsBankDetailsUpdated?: string;
  emailId?: string;
  mobileNumber?: string;
  accountId?: string;
  accountCreationDate?: string;
  lastUpdateTimestamp?: string;
}